# Performance improvement of KFunction.callBy