@k163377 2022-10-22
@udalov 
Would it be difficult for you to review this PR?
I think the `KFunction.callBy` speedup is an important change for `kotlin-reflect` users.

If you are concerned about the amount of changes or increased memory consumption due to the introduction of cache, I can change the content as follows
https://github.com/Kotlin/kotlinx.reflect.lite/pull/36