@udalov 2022-11-01
Sorry for the delay! I agree, this optimization can be very useful.


I think we can go further and make the `getAbsentArguments` array of size `parameters.size + (if (isSuspend) 1 else 0) + masks.size + 1`. This will get rid of another array creation at the end of `callDefaultMethod`. This will, however, require to copy the array in case no default arguments were omitted (to take its prefix of size `parameters.size + (if (isSuspend) 1 else 0)`), which is the `!anyOptional` case. But maybe that's fine since the user can also use the more performant `KCallable.call` in that case anyway. What do you think?