@k163377 2022-11-02
Thanks for the review.

I agree with your suggestions.
I think it is more reasonable to optimize `callBy` to use default values, since the optimization of using `call` when all arguments are given can be easily done by the user.

I will revise and push this when I have time.