@udalov 2022-11-08
Thanks! I've also noticed that now we perform integer boxing when computing the mask for every default argument and I'd like to avoid it. What do you think about the following (hopefully last) optimization https://github.com/JetBrains/kotlin/commit/388fdc4ae54ae68b383ffa78c9cd5e1a882223f2?