@k163377 2022-11-11
> What do you think about the following (hopefully last) optimization https://github.com/JetBrains/kotlin/commit/388fdc4ae54ae68b383ffa78c9cd5e1a882223f2?

It does not seem to be a good fix, at least as far as the benchmark results are concerned.

Even with this optimization, the speedup appears to be only when many default arguments are used.
In other cases, it appears to be offset by the cost of branching.

**[638d89c](https://github.com/JetBrains/kotlin/pull/4840/commits/638d89c033e2d6312fc3884c27dd1715dc118eae)**

```
Benchmark                          Mode  Cnt         Score        Error  Units
Measurement.zero                  thrpt    4  12938281.870 ± 488259.178  ops/s
Measurement.fiveWithDefault       thrpt    4   1940327.443 ± 198156.544  ops/s
Measurement.fiveWithoutDefault    thrpt    4   2442313.516 ±  39665.853  ops/s
Measurement.oneWithDefault        thrpt    4   5451979.125 ± 715952.574  ops/s
Measurement.oneWithoutDefault     thrpt    4   6475176.599 ± 349258.846  ops/s
Measurement.twentyWithDefault     thrpt    4    505560.683 ±  42324.766  ops/s
Measurement.twentyWithoutDefault  thrpt    4    726913.081 ± 136844.449  ops/s
```

**When https://github.com/JetBrains/kotlin/commit/388fdc4ae54ae68b383ffa78c9cd5e1a882223f2 is imported**

```
Benchmark                          Mode  Cnt         Score         Error  Units
Measurement.zero                  thrpt    4  13029498.008 ±  255137.936  ops/s
Measurement.fiveWithDefault       thrpt    4   1832503.215 ± 1117903.663  ops/s
Measurement.fiveWithoutDefault    thrpt    4   2399084.679 ±  117597.661  ops/s
Measurement.oneWithDefault        thrpt    4   5460558.554 ±  279303.773  ops/s
Measurement.oneWithoutDefault     thrpt    4   6325791.468 ±  187765.117  ops/s
Measurement.twentyWithDefault     thrpt    4    598262.914 ±   26892.919  ops/s
Measurement.twentyWithoutDefault  thrpt    4    729748.383 ±  100592.024  ops/s
```