@udalov 2022-11-29
I've merged all commits and slightly changed comments in https://github.com/JetBrains/kotlin/commit/3393bb40a4f38820d5813122b576e2a6fc5b3612. I'll push it to master after the build succeeds.