@udalov 2022-11-29
Merged to master. This change will be released as a part of Kotlin 1.8.20.

Thank you for the contribution! And in particular, thanks for extensive measurement of effects of all the discussed changes. :+1: